Accomodate for Versioned extension
Problem: There isn't any implementation for DataObjects using Versioned extension. When we retrieve the DataObjects during the result search, it causes a blank screen as it can't resolve 'staging' models. 

Solution: Check if the model is using the Versioned extension then grab the 'Live' DataObjects to pull into the table instead.

Extra bonus: Bug with the original code as you had two declarations of $dos with the latter overriding the prior. It does not pull in the getSearchFilter function into the filter (uses empty array instead).